2 more image comparison methods
I have added two new methods for image comparison: Peak Signal to Noise Ratio Comparison and Histogram Comparison. If there is a match of 2 or more between the old and the new methods, the matching hero is selected. If there is no specific match, then only the subtract method match is taken into account.